feat: support for chat and OpenAPI services
We now provides Synths with better encapsulation. Instead of one big
Synth with the implementation of the whole system, a Synth can have
just the code it needs to implement a particular component. This
complicates the design in some ways, but also makes it easier to have
isolated (automated) experiments that can't break other parts of the
system.

Add an experiment `dump` action to dump a transitive closure of units

Also fix some bugs in dependency detection for expressions
like `export { foo } from "bar"`