feat: late-binding constraints and capture
We now treat standalone comparisons (including ones with arithmetic
expressions) as late-binding constraints.

If referenced variables are bound to concrete values when the
constraint declaration is encountered, then the constraint is applied
immediately and existing semantics continue to hold (the proof path is
pruned if the constraint fails, continues if it passes).

If referenced variables are not yet bound, the constraint is attached
to the variable entry and will be checked when that variable is bound,
which might result in it being immediately pruned.

This means that they are tested only when all variables referenced in
the expression are bound to concrete values.

Note that this does not include expressions wrapped by a call to
Logic.js. Logic.js(...) expressions are still eagerly evaluated as
before.

Late-binding constraints means we can use the aggregated network of
constraints to find solutions.

We do not yet define any built-in constraint solvers. Instead, we
provide the Logic.constraints(...) built-in, which retrieves
constraints associated with a variable.