fix: make _ an anonymous var in objects/arrays
When _ is used directly in a variable we can safely ignore the binding,
but when used in an object or array it represents a specific hole that
might be bound later. We need a new anonymous symbol in these cases.

Adjust test previously wrong test assertions accordingly.

Also, use $objectPattern and $arrayPattern instead of $class encoding to
decouple how we write test from the runtime contract of these classes.